// src/model/ProjectData.tsx
interface ProjectData {
  id: number;
  name: string;
  description: string;
  image: string;
  link: string;
  tags: string[];
  responsibilities: string[];
  technologies: string[];
  status: string;
  supportedChains: string[];
  partners: string[];
  keyContributions: string[];
}

export default ProjectData;